refactor(platform): extract build_sdk_env to dedicated module
Move build_sdk_env() from copilot/sdk/service.py to copilot/sdk/env.py
to break the circular import chain:
  tool_orchestrator → copilot.sdk.service → executor → credit → block_cost_config → tool_orchestrator

The new env.py module only imports copilot.config and subscription,
avoiding the heavy service.py import chain. service.py re-exports
build_sdk_env for backward compatibility.
